add firestore tests - refactor and ts-ifying proved necessary to do this
