+++ src/init/features/functions/index.js
@@ -7,7 +7,7 @@ var _ = require("lodash");
 var logger = require("../../../logger");
 var { prompt } = require("../../../prompt");
 var enableApi = require("../../../ensureApiEnabled").enable;
-var requireAccess = require("../../../requireAccess");
+var requireAccess = require("../../../requireAccess").requireAccess;
 var scopes = require("../../../scopes");
 
 module.exports = function(setup, config) {
