+++ src/init/features/index.js
@@ -2,7 +2,7 @@
 
 module.exports = {
   database: require("./database"),
-  firestore: require("./firestore").doSetup,
+  firestore: require("./firestore/firestore").doSetup,
   functions: require("./functions"),
   hosting: require("./hosting"),
   storage: require("./storage").doSetup,
